backend/requirements.txt
# Backend Dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
httpx>=0.25.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
pydantic>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0